Printlab12

+++ AbstractPrinteryView/FormCreateEdition.cs
@@ -113,7 +113,7 @@ namespace AbstractPrinteryView
                     CustomerNumber = Convert.ToInt32(comboBoxCustomer.SelectedValue),
                     EditionNumber = Convert.ToInt32(comboBoxEdition.SelectedValue),
                     Count = Convert.ToInt32(textBoxCount.Text),
-                    Sum = Convert.ToInt32(textBoxSum.Text)
+                    Sum = Convert.ToDecimal(textBoxSum.Text)
                 });
                 MessageBox.Show("Сохранение прошло успешно", "Сообщение", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 DialogResult = DialogResult.OK;
